split Iterator.h into 2 separate headers Removed the header forward declaration
